Precision changed from 1 ms to 0.1 ms.

+++ Arduino/LagMeter/src/Measurement/Measure.h
@@ -3,8 +3,8 @@
 
 // Structure to return min/max values.
 struct MinMax {
-  int min;
-  int max;
+  int16_t min;
+  int16_t max;
 };
